update action and enum action var

+++ MyLoginScreen/app/src/main/java/com/july/myloginscreen/ContactsRepository.kt
@@ -7,26 +7,32 @@ import androidx.lifecycle.MutableLiveData
 
 
 /**
- * Esta clase hace uso de la interfaz NotaDao e interactúa con la base de datos.
+ * Esta clase hace uso de la interfaz ContactoDao e interactúa con la base de datos.
  *
- * @author Julio Maker
+ * @author July Maker
  * @version 2020.00
  */
 
 class ContactsRepository(application: Application) {
 
+    enum class ACTION{
+        INSERTAR,
+        ACTUALIZAR,
+        BORRAR
+    }
+
     private val contactDao: ContactDao? = ContactsDatabase.getInstance(application)?.contactDao()
 
     fun insert(contact: Contact) {
-        if (contactDao != null) InsertAsyncTask(contactDao).execute(contact)
+        if (contactDao != null) ActionAsyncTask(contactDao, ACTION.INSERTAR).execute(contact)
     }
 
     fun update(contact: Contact) {
-        if (contactDao != null) contactDao.update(contact)
+        if (contactDao != null) ActionAsyncTask(contactDao, ACTION.ACTUALIZAR).execute(contact)
     }
 
     fun delete(contact: Contact) {
-        if (contactDao != null) DeleteAsyncTask(contactDao).execute(contact)
+        if (contactDao != null) ActionAsyncTask(contactDao, ACTION.BORRAR).execute(contact)
     }
 
 
@@ -34,21 +40,22 @@ class ContactsRepository(application: Application) {
         return contactDao?.getOrderedAgenda() ?: MutableLiveData<List<Contact>>()
     }
 
-    private class InsertAsyncTask(private val contactDao: ContactDao) :
+    private class ActionAsyncTask(private val contactDao: ContactDao, private val action:ACTION) :
         AsyncTask<Contact, Void, Void>() {
         override fun doInBackground(vararg contacts: Contact?): Void? {
             for (contact in contacts) {
-                if (contact != null) contactDao.insert(contact)
-            }
-            return null
-        }
-    }
+                when(action){
+                    ACTION.INSERTAR->{
+                        if (contact != null) contactDao.insert(contact)
+                    }
+                    ACTION.ACTUALIZAR->{
+                        if (contact != null) contactDao.update(contact)
+                    }
+                    ACTION.BORRAR->{
+                        if (contact != null) contactDao.delete(contact)
+                    }
+                }
 
-    private class DeleteAsyncTask(private val contactDao: ContactDao) :
-        AsyncTask<Contact, Void, Void>() {
-        override fun doInBackground(vararg contacts: Contact?): Void? {
-            for (contact in contacts) {
-                if (contact != null) contactDao.delete(contact)
             }
             return null
         }
